Refactor price - Float->Long

+++ app/models/Product.scala
@@ -3,7 +3,7 @@ package models
 import play.api.libs.json.Json
 import play.api.libs.json.OFormat
 
-case class Product(id: Int, name: String, description: String, category_id: Int, price: Float)
+case class Product(id: Int, name: String, description: String, category_id: Int, price: Long)
 
 object Product {
   implicit val productFormat: OFormat[Product] = Json.format[Product]
